+++ CHANGELOG.md
@@ -6,6 +6,9 @@ All notable changes to [@bpmn-io/variable-resolver](https://github.com/bpmn-io/v
 
 ___Note:__ Yet to be released changes appear here._
 
+* `FIX`: show FEEL expression for unresolvable function invocations ([#104](https://github.com/bpmn-io/variable-resolver/pull/104), [camunda/camunda-modeler#5744](https://github.com/camunda/camunda-modeler/issues/5744))
+* `DEPS`: update to `@bpmn-io/lezer-feel@2.4.0`
+
 ## 3.0.0
 
 * `FEAT`: expose both read and written variables. `getVariablesForElement` can be used to retrieve variables for a specific element ([#74](https://github.com/bpmn-io/variable-resolver/pull/74))